This is a bulk update of the providers code that handles the needs of being installed EL8-based systems.
There is a base change for the /mirrors1 save path to my opinionated /repos path.
There is a base change for bandersnatch where bandersnatch was trying a pip install from reference
  that is now replaced with a direct call to pip install bandersnatch.
There are a few python hacks to set python3 as the bin name.
Removed some of the ruby settings and moved them to the base install of singularity-base scripts
  which is clearly not ideal as it makes this code dependent (without hand alterations) on the
  singularity bits, but realistically I'm going to be the only person who ever uses this software
  but if you are reading this it means you and I are using it.
There is a bit of linting removing trailing spaces in files.
Also altered the wrapper for reposync which is now under dnf which has removed the -r option from
  the command arguments and replaced with a hard requirement for --repoid.
A bit of alterations to gem.